backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.3
fastembed>=0.3.0
chromadb>=0.6.0
duckdb>=0.10.3
psycopg2-binary>=2.9.9
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv==1.0.1
pydantic==2.8.0
numpy>=1.24.0
scipy>=1.11.0
tenacity==8.3.0
